Create Tolkien Week 7 Show and Tell
My computer is having trouble working with M2 right now so I'm submitting it like this.  This is the code I used for the group project on Friday.  Some of the issues I've been having are:  I'm not sure if I'm doing the right thing for returning the generators of the groebner basis since it just seems to output the ones I used in the first place maybe its an issue with my monomial ordering, though.  quotientRemainder is the wrong method to be using here but I haven't been able to figure out the right one for multivariable division.